Added a step to my workflow file
I used GitHub CLI (gh) to add a comment when a pull request is opened. To allow GitHub CLI to post a comment, I set the GITHUB_TOKEN environment variable to the value of the GITHUB_TOKEN secret, which is created when the workflow runs. I also set the PR_URL environment variable to the URL of the newly created pull request.